clear code, add plug of double requerst update method

+++ src/model/EntityUser.java
@@ -24,7 +24,7 @@ public class EntityUser implements Serializable {
         return name;
     }
 
-    public int getAge() {
+    public byte getAge() {
         return age;
     }
